UserInfo.js - Validation
Validation - has empty object for error and whenever condition does not match to the input the error will be pushed in that object. at the end this object will be returned from function.

on submit  function - handleSubmit will be called. it will call validate function and the returned object will be passed to errors state. than if errors state is empty ,  page will navigate to second registration page - laptopinfo.