LaptopInfo.js - Validation
Validation - has empty object for error and whenever condition does not match to the input the error will be pushed in that object. at the end this object will be returned from function.

on submit function - handleSubmit will be called. it will call validate function and the returned object will be passed to errors state. than getFormData loops through the data array and for each input value appends each key and value to FormData object. this FormData object will be send to Post Api call.

than on submit button  if errors state is empty form will be submited. api call will be made, localstorage will be cleared and we will navigate on lastpage(congrats page-popup)